Added score + happiness lvl
Patients now have a happiness level (not displayed on screen) that decreases over time. Game score gets increased when a patient has been healed.

+++ Hospital Saviour/Assets/Scripts/Sickness.cs
@@ -8,5 +8,5 @@ public class Sickness : ScriptableObject
 {
     public SicknessBase sicknessBase;
     public List<Sprite> healingOrderIcons;
-    public float happinessDropLevel;
+    public int happinessDropLevel;
 }
